requirements.txt
apscheduler==3.10.4
flask==3.1.0
flask-sqlalchemy==3.1.1
gunicorn==23.0.0
itsdangerous==2.2.0
jinja2==3.1.4
openpyxl==3.1.5
pandas==2.2.3
Pillow==11.1.0
PyJWT==2.9.0
python-dateutil==2.9.0
reportlab==4.2.5
requests==2.32.3
sqlalchemy==2.0.36
werkzeug==3.1.3
xlrd==2.0.1
python-dotenv==1.0.1
xlwt==1.3.0
flask-cors==5.0.0
twilio==9.4.0
